+++ internal/backend/remote-state/gcs/client.go
@@ -7,6 +7,7 @@ package gcs
 
 import (
 	"encoding/json"
+	"errors"
 	"fmt"
 	"io"
 	"strconv"
@@ -135,7 +136,10 @@ func (c *remoteClient) lockError(err error) *statemgr.LockError {
 	}
 
 	info, infoErr := c.lockInfo()
-	if infoErr != nil {
+	if errors.Is(infoErr, storage.ErrObjectNotExist) {
+		// Race condition - file exists initially but then has been deleted by other process
+		lockErr.InconsistentRead = true
+	} else if infoErr != nil {
 		lockErr.Err = multierror.Append(lockErr.Err, infoErr)
 	} else {
 		lockErr.Info = info